+++ Sources/Picker.cpp
@@ -7,13 +7,13 @@ Picker::~Picker()
 {
 }
 
-void Picker::calcualteMouseRay(int width, int height, 
+void Picker::calcualteMouseRay(float width, float height, 
 							   int mouse_x, int mouse_y, glm::mat4 view)
 {	
 	glm::vec2 ray_nds = getNDS(width, height, mouse_x, mouse_y); // Screen space to Normalized Device space
 	glm::vec4 ray_clip = glm::vec4(ray_nds, -1.0, 1.0); // ND Space to Clipped space
 	
-	float aspect = static_cast<float>(width) / static_cast<float>(height);
+	float aspect = width / height;
 	glm::vec4 ray_camera = getCameraSpace(ray_clip, aspect);
 	m_ray_world = getWorldSpace(ray_camera, view);
 }
